refactor: ERROR_MESSAGES 객체에서 maxLengthNumber 속성을 정적 문자열로 변경
