+++ hooks/src/lib/constants/messages.ts
@@ -4,6 +4,5 @@ export const ERROR_MESSAGES = {
   expired: "유효 기간이 만료되었습니다. 유효한 기간을 입력해 주세요.",
   company: "카드사를 선택해 주세요.",
   holder: "유효하지 않은 이름입니다.",
-  maxLengthNumber: (maxLength: number) =>
-    `${maxLength}자리 숫자로 입력해 주세요.`,
+  maxLengthNumber: "자리 숫자로 입력해 주세요.",
 };
